[MCUX-66405] Change default key store ID to 1
The S200 only support one keystore, so we were using 0 as a hardcoded
default. This would however cause repeated calls to fail unnecessarily,
as well as cause some issues with mutlicore multiple-initialization from
each core.

We now use ID 1 instead, as we may call the API repeatedly. In such a
case we transparently pass without any change to the key store.

Signed-off-by: Patrik Nemeth